requirements.txt
Pillow
transformers
spacy
torch 